Move maybeMergeOptions into CommandLineUtils.
